pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "overture-pipeline"
version = "0.1.0"
description = "Cloud-native ETL pipeline for Overture Maps data to ArcGIS Online"
readme = "README.md"
requires-python = ">=3.11"

authors = [
    { name = "Ryan Arias Delafosse, World Bank", email = "rariasdelafosse@worldbank.org" }
]

dependencies = [
    # Core geospatial and data processing
    "arcgis>=2.4,<2.5",
    "arcgis-mapping",
    "duckdb>=1.1.3,<1.2",      
    "pyarrow>=16.1.0,<17.0",   # Constrained by ArcGIS API requirements
    "geopandas>=1.0.1,<1.1",   
    "pyogrio>=0.10.0,<0.11",   
    "shapely>=2.0.6,<2.1",     
    "pyproj>=3.6.1,<3.7",
    "pandas>=2.0.0,<2.2.0",    # Constrained by ArcGIS API requirements
    
    # HTTP and cloud access for optimized downloads
    "requests>=2.31.0,<3.0",
    "boto3>=1.34.0,<1.35",     # Optional S3 optimization
    "tqdm>=4.66.0,<5.0",       # Progress bars for downloads
    
    # CLI and configuration
    "typer==0.12.3",
    "click<8.2.0",  # Typer compatibility constraint
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0,<3.0",  # Type validation and domain models
    
    # Development tools
    "pytest>=8.2.0",
    "ruff>=0.5.6",
    "mypy>=1.10.0",
]

[project.scripts]
o2agol = "o2agol.cli:app"

[tool.ruff]
line-length = 100
target-version = "py311"
lint.select = ["E","F","I","UP","B","SIM"]
lint.ignore = ["E501", "UP045"]  # UP045: Typer doesn't support X | None syntax

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
